Extract Address extension shared methods for watch target
